pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arkprobe"
version = "0.1.0"
description = "ArkProbe - Workload characterization framework for chip design space exploration on Kunpeng ARM processors"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "rich>=13.0",
    "plotly>=5.18",
    "scikit-learn>=1.3",
    "kaleido>=0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[project.scripts]
arkprobe = "arkprobe.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["arkprobe*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
